feat: rename Accounting\AccountingOwnerInterface
